Moved JobStatusResponse.cs to appropriate namespace
